// src/contacts/models/contact.ts
export class Contact {
  id?: string;
  cell?: string;
  city?: string;
  country?: string;
  dob?: Date;
  email?: string;
  firstName?: string;
  fullTextSearch?: string;
  imageUrl?: string;
  lastName?: string;
  phone?: string;
  postalCode?: string;
  state?: string;
  street?: string;
  thumbnailUrl?: string;
  timezone?: string;
}